Add 'git-gerrit-push' script
Add a script implementing our standard "git gerrit-push" alias.  When
invoked, first check that the current branch is a named topic.  Then
fetch 'master' from the 'gerrit' remote repository so the local
repository can compute a minimal pack to push.  Finally, push the topic
to the 'gerrit' remote as "refs/for/master/$topic".