Updated space_parse with a noLeadingSpace flag
There is now a flag that tells the space_parse method to enable or
disable the omit a leading space option.  Be default the flag is true,
i.e. leading spaces are omitted.

In the case of a "See" node, the flag is set to False before the
nodes children are traversed.  Afterwards, the flag is set back
to True.  The effect is that children of the See node do get a
leading space, indenting them 1 space.

Change-Id: I1d0a92c2231705d8d1e40a4b67acb9eff6b9737d