BUG: Fix parameter map test broken by previous commit
A previous commit forced parameter map writer to always use floats in
order to avoid scientific notation. This broke a test that compared a
parameter map with the same parameter map written and read from disk
because ints had been converted to float. In the new test we instead
use the two different parameter maps to run registrations and compare
the hash of resulting images.