Don't use ITK_MODULES_ENABLED for complete list of install ITK modules
The ITK_MODULES_ENABLED variable may only include ITK modules enabled
during the initial configuration, and may not include install remote
or external modules.

It is preferred to use find_packages to "query" if ITK has a module
available, as a complete list of install ITK module may not exist.

Suggested-by: Matt McCormick <matt.mccormick@kitware.com>