Must use a ImageSeriesReader to read a DICOM series
The test DICOM series, does not have an identity transform. The series
contains a per-slice y change in addition to the z-spacing
change. The join series filter expects that the x and y coordinates
are the same, so for this data-set an exception is generated.

The example has been corrected to use the image series reader to
properly read the volume, then also use the image file reader to
re-read each slice to obtain the meta-data dictionary.