Use Rscript instead of the R executable
Using the Rscript command is preference in the R community over using
the R executable with the `--file` argument. Switching to Rscript
removes extra output of the R executable including output to prompt a
user for an interactive sessions, and the echoing of the input
file. The `--file` and `--args` flags are no longer needed when using
sitk_add_r_test, making the usage similar to that of other languages.