Only warn when the source version is not available.
The SimpleITK version expects to be able to determine the version from
a git repository or from saved information from the repository. This
patch not produces a only a warning when the version information is
not properly available. The version will fallback to a non specific
"dev" without a number. This will test less thank any other version,
since it's not a proper version of SimpleITK.