Updating the documentation of the FilterProgressReporting example.
The documentation is a duplicate of the Doxygen page with some minor
modifications. The doxygen markup is also used in the sphinx
documentation. Before we can remove the doxygen page we need to
figure out how to link between the doxygen generated pages to the
sphinx generated pages on readthedocs. There are references from some
classes to the documentation in the sphinx/doxygen example page.

Change-Id: Ic15f4e6f51f3079a4ade058bfe4003fccbe1f599