Replace configuration_file_build_time
The old build time configuration file was fundamentally flawed. The
configuration specific path to a target is can be only conveyed with
CMake via generator expressions. But this function did not use them,
but relied on setting and the LOCATION property of the target to the
same location for each target. So the build time configuration wasn't
even needed. Also the LOCATION manipulation produced a CMP0026
warning.

This new approach does an initial basic configuration, then does a
second pass which actually does a build time configuration with CMake
generator expressions manually passed.

The old CMake file to provide the function has been removed.