Change to vector argument for sigma and variance of Gaussian
Enable setting of the sigmas/variance on a per axis basis for the
SmoothingRecursiveGaussianImageFilter and the
DiscreteGaussianImageFilter. This is done may changing the parameter
to be a "dim_vec". The scalar method for the object oriented interface
remains.

Additionally, the signature of scalar procedural remains as part of
the Additional procedural interface.