Relax the "short" image testing tolerance for discrete Gaussian
On the i386 Debian build, it has been reported that the
BasicFilters.DiscreteGaussianImageFilter_short test is failing with a
RMS difference of 0.513965. This tolerance has already been relaxed
once, and further relaxing seems reasonable due to intermediate image
of float being converted to short without proper rounding.

While this is an indication the internal ITK algorithm could be
improved for portability and accuracy in ITK, it is not an indication
of a problem with the SimpleITK build.