Provide instructions for building readthedocs documentation in a Python environment
The current documentation explains how to install the Python Sphinx in a Debian-based platform and run it via a Makefile, but some users may want to run it directly in a Python environment.This documentation update provides instructions about installing the appropriate Sphinx package and theme in a Python environment and running Sphinx directly to generate the documentation. This method also bypasses the current Makefile's complaint about not finding a source directory.