ENH: Add support for deformation field
Previously transformix would write the deformation field to disk. A
recent improvement to elastix exposed the deformation field via the
library API, and the writing to disk was not done anymore, which was
unfortunate for SimpleElastix users who depended on the def field being
written to disk. This commit adds GetDeformationField to transformix so
deformation field can be retrieved directly from memory. Fixes #181 and
fixes #182.